requirements.txt
pandas
numpy
requests
python-dotenv
plotly
streamlit
yfinance
mplfinance
pytest
